+++ haha/handlers/summaries.go
@@ -4,10 +4,10 @@ import (
 	"encoding/json"
 	"github.com/gorilla/mux"
 	"io/ioutil"
+	"joblessness/haha/models"
+	"joblessness/haha/utils/pdf"
 	"log"
 	"net/http"
-	"server/server/models"
-	"server/server/utils/pdf"
 	"strconv"
 	"sync"
 	"sync/atomic"
