+++ haha/models/summary.go
