+++ haha/models/vacancy.go
