+++ haha/tests/users_test.go
@@ -4,10 +4,10 @@ import (
 	"bytes"
 	"encoding/json"
 	"github.com/gorilla/mux"
+	"joblessness/haha/handlers"
+	"joblessness/haha/models"
 	"net/http"
 	"net/http/httptest"
-	"server/server/handlers"
-	"server/server/models"
 	"sync"
 	"testing"
 	"time"
