+++ haha/tests/vacancies_test.go
@@ -4,10 +4,10 @@ import (
 	"bytes"
 	"encoding/json"
 	"github.com/gorilla/mux"
+	"joblessness/haha/handlers"
+	"joblessness/haha/models"
 	"net/http"
 	"net/http/httptest"
-	"server/server/handlers"
-	"server/server/models"
 	"strings"
 	"sync"
 	"testing"
